chore: refactor database detail
